expr resolve functions return a replacement expr
if the expression fails to resolve, resolve functions return NULL.

If the expression pointer needs to be replaced (i.e. for optimizations or for when an expression must be reinterpreted as a different type of expression) it returns a pointer to a new expression.

otherwise, the original expression is returned.